ui/requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
pyarrow>=12.0.1
openpyxl>=3.1.2
